scripts/requirements.txt
google-generativeai==0.8.3
python-dotenv==1.0.1
pytest==8.2.0
